refactor: countBallsAndStrikes, generateAnswer, userInputValidation 함수 모듈화
