+++ README.md
@@ -35,12 +35,13 @@
 ### 4. 리팩토링 - 작성했던 함수 모듈화하기
 
 - [x] `compareAnswer` 함수를 `countBallsAndStrikes`와 `showBallsAndStrikes` 함수로 분리하기
-- [] 만들었던 각 함수를 `utils` 디렉토리 내부에서 `import`로 가져와서 사용하기
+- [x] 만들었던 각 함수를 `utils` 디렉토리 내부에서 `import`로 가져와서 사용하기
 - [x] 로직에 적합한 변수명, 함수명 사용하기
 
 ### 5. 버그
 
 - [x] 재시작 시 컴퓨터가 생성한 답안이 재생성되지 않던 버그 수정하기
+- [x] cypress 테스트 결과값 도출을 위해 `play` 메서드에 반환값 추가
 
 ### 테스트
