+++ package.json
@@ -6,7 +6,7 @@
     "@woowacourse/mission-utils": "^1.0.1"
   },
   "devDependencies": {
-    "cypress": "8.7.0"
+    "cypress": "^8.7.0"
   },
   "scripts": {
     "test": "cypress run --browser chrome"
